requirements.txt
Django==1.10.2
django-post-office==2.0.8
httplib2==0.10.3
jsonfield==1.0.3
jsonpickle==0.9.4
oauth2==1.9.0.post1
oauth2client==4.0.0
Pillow==3.4.2
pyasn1==0.2.2
pyasn1-modules==0.0.8
rsa==3.4.2
six==1.10.0
